moved functions back to seniorDesign
